FIX #24 - Organization - Settings
 - update organization - settings

+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/services/UserOrganizationService.java
@@ -111,7 +111,37 @@ public class UserOrganizationService {
         return userOrganization;
     }
 
+
+    public UserOrganization updateUserRole(UUID userId, UUID organizationId, String roleName) throws GendoxException {
+        UserOrganization userOrganization = userOrganizationRepository.findByUserIdAndOrganizationId(userId, organizationId);
+        if (userOrganization == null) {
+            throw new GendoxException("USER_ORGANIZATION_NOT_FOUND", "User-organization combination not found", HttpStatus.BAD_REQUEST);
+        }
+        Type role = typeService.getOrganizationRolesByName(roleName);
+        userOrganization.setRole(role);
+        return userOrganizationRepository.save(userOrganization);
+    }
+
+    public UserOrganization updateUserRole(UUID userOrganizationId, String roleName) throws GendoxException {
+        UserOrganization userOrganization = userOrganizationRepository.findById(userOrganizationId).orElse(null);
+        if (userOrganization == null) {
+            throw new GendoxException("USER_ORGANIZATION_NOT_FOUND", "User-organization combination not found", HttpStatus.BAD_REQUEST);
+        }
+        Type role = typeService.getOrganizationRolesByName(roleName);
+        userOrganization.setRole(role);
+        return userOrganizationRepository.save(userOrganization);
+    }
+
+
     public void deleteUserOrganization(UserOrganization userOrganization) throws GendoxException {
+        userOrganizationRepository.delete(userOrganization);
+    }
+
+    public void deleteUserOrganization(UUID organizationId, UUID userId) throws GendoxException {
+        UserOrganization userOrganization = userOrganizationRepository.findByUserIdAndOrganizationId(organizationId, userId);
+        if (userOrganization == null) {
+            throw new GendoxException("USER_ORGANIZATION_NOT_FOUND", "User-organization combination not found", HttpStatus.BAD_REQUEST);
+        }
         userOrganizationRepository.delete(userOrganization);
     }
 }
